修正３

+++ app/controllers/items_controller.rb
@@ -22,10 +22,14 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if user_signed_in?
+    if user_signed_in? 
       item = Item.find(params[:id])
-      item.destroy
-      redirect_to root_path
+      if current_user.id == item.user.id
+        item.destroy
+        redirect_to root_path
+      else
+        render :index
+      end
     end
   end
